requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
codecov==2.1.9
colorama==0.4.3
coverage==5.3
flake8==3.8.4
google-api-core==1.22.2
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
iniconfig==1.0.1
mccabe==0.6.1
multidict==4.7.6
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.6.0
